authentication fixes
