+++ Middleware/src/main/java/ca/concordia/encs/citydata/core/AuthController.java
@@ -1,0 +1,39 @@
+package ca.concordia.encs.citydata.core;
+
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.ResponseEntity;
+import org.springframework.security.authentication.AuthenticationManager;
+import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
+import org.springframework.security.core.Authentication;
+import org.springframework.security.core.AuthenticationException;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RestController;
+
+import ca.concordia.encs.citydata.authentication.AuthRequest;
+import ca.concordia.encs.citydata.authentication.AuthResponse;
+import ca.concordia.encs.citydata.authentication.JwtUtil;
+
+@RestController
+public class AuthController {
+
+	@Autowired
+	private AuthenticationManager authenticationManager;
+
+	@Autowired
+	private JwtUtil jwtUtil;
+
+	@PostMapping("/authenticate")
+	public ResponseEntity<?> createToken(@RequestBody AuthRequest authRequest) {
+		try {
+			Authentication authentication = authenticationManager.authenticate(
+					new UsernamePasswordAuthenticationToken(authRequest.getUsername(), authRequest.getPassword()));
+
+			String jwt = jwtUtil.generateToken(authRequest.getUsername());
+			return ResponseEntity.ok(new AuthResponse(jwt));
+		} catch (AuthenticationException e) {
+			return ResponseEntity.status(401).body("Invalid username or password");
+		}
+	}
+
+}
