+++ Middleware/src/main/java/ca/concordia/encs/citydata/core/controllers/PublicController.java
@@ -1,0 +1,15 @@
+package ca.concordia.encs.citydata.core.controllers;
+
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
+
+@RestController
+@RequestMapping("/public")
+public class PublicController {
+
+	@GetMapping("/ping")
+	public String ping() {
+		return "Service is alive";
+	}
+}
