+++ Middleware/src/main/resources/application.properties
@@ -6,3 +6,9 @@ spring.application.name=CityData
 mongo.enabled=false
 spring.data.mongodb.uri=mongodb://localhost:27017/citydata
 spring.autoconfigure.exclude=org.springframework.boot.autoconfigure.mongo.MongoAutoConfiguration,org.springframework.boot.autoconfigure.data.mongo.MongoDataAutoConfiguration
+
+# Enable detailed security logs
+logging.level.org.springframework.security=DEBUG
+
+# Ensure default security configuration is not auto-wired
+spring.main.allow-bean-definition-overriding=true
